requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord.py==2.3.2
Flask==3.0.3
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.1
